+++ tests/test_models/test_base.py
@@ -156,3 +156,8 @@ def test_deep_base_model_forecast_loop(simple_df, deep_base_model_mock):
     np.testing.assert_allclose(
         future.df.loc[:, pd.IndexSlice["B", "target"]], raw_predict[("B", "target")][:horizon, 0]
     )
+
+
+def test_deep_base_model_forecast_throw_error_on_return_components():
+    with pytest.raises(NotImplementedError, match="This mode isn't currently implemented!"):
+        DeepBaseModel.forecast(self=Mock(), ts=Mock(), prediction_size=Mock(), return_components=True)